App restructured
