+++ Android/app/src/main/java/com/bazukaa/secured/authentication/FingerprintHandler.java
@@ -1,4 +1,4 @@
-package com.bazukaa.secured.ui;
+package com.bazukaa.secured.authentication;
 
 import android.app.Activity;
 import android.content.Context;
